Rename UploadPage file to UploadSection
